backend/langchain/requirements.txt
flask
flask-restful
flask-jwt-extended
pika  
python-dotenv  
uvicorn
langchain
requests
pillow
flask-cors